Not sure if my previous commit for this went through, but added the changes made for that commit:
Added logMovement(int x, int y) and logMovement(Tile t)  - they do the same thing... one is getting the Tile as the parameter, the other uses the x and y coordinates.  Also, added the filename to the "New file created:" when a new logging file is created, just to keep track of file creation.  WriteCurrentLocationToFile java class no longer needed with the addition of these methods.